pyproject.toml
[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[project]
name = "gratekeeper"
version = "0.4.0"
description = "GRatekeeper is a lightweight GitHub API helper with proactive rate-limit throttling"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.31.0",
    "rich>=13.7.0",
    "textual>=0.58.0",
]
authors = [{ name = "Really Him", email = "hesreallyhim@proton.me" }]
license = { text = "MIT" }

[project.scripts]
gratekeeper-dashboard = "gratekeeper.dashboard:main"
gk-dash = "gratekeeper.dashboard:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pre-commit>=3.7.0",
    "types-requests>=2.31.0.10",
]


[tool.hatch.build.targets.wheel]
packages = ["src/gratekeeper"]

[tool.pytest.ini_options]
markers = [
    "demo: live GitHub demo scenarios (requires GRATEKEEPER_RUN_DEMOS=1)",
]
